+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/pages/employee/SetupCampaign.js
@@ -62,7 +62,7 @@ const SetupCampaign = ({ hierarchyType, hierarchyData }) => {
   const [active, setActive] = useState(0);
   const { data: HierarchySchema } = Digit.Hooks.useCustomMDMS(tenantId, CONSOLE_MDMS_MODULENAME, [{ 
     name: "HierarchySchema",
-    "filter": "[?(@.type=='console')]"
+    "filter": `[?(@.type==${window.Digit.Utils.campaign.getModuleName()})]`
    }],{select:(MdmsRes)=>MdmsRes},
    { schemaCode: `${CONSOLE_MDMS_MODULENAME}.HierarchySchema` });
   const lowestHierarchy = useMemo(() => {