+++ health/micro-ui/web/micro-ui-internals/packages/modules/microplan/src/configs/UICustomizations.js
@@ -123,7 +123,7 @@ export const UICustomizations = {
             <div>
               {microplanFileId && row?.status == "RESOURCE_ESTIMATIONS_APPROVED" ? (
                 <div>
-                  <ButtonNew style={{ width: "20rem" }} icon="DownloadIcon" onClick={handleDownload} label={t("WBH_DOWNLOAD_MICROPLAN")} />
+                  <ButtonNew style={{ width: "20rem" }} icon="DownloadIcon" onClick={handleDownload} label={t("WBH_DOWNLOAD_MICROPLAN")} title={t("WBH_DOWNLOAD_MICROPLAN")}  />
                 </div>
               ) : (
                 <div className={"action-button-open-microplan"}>
@@ -132,6 +132,7 @@ export const UICustomizations = {
                       type="actionButton"
                       variation="secondary"
                       label={t("MP_ACTIONS_FOR_MICROPLAN_SEARCH")}
+                      title={t("MP_ACTIONS_FOR_MICROPLAN_SEARCH")}
                       options={options}
                       style={{ width: "20rem" }}
                       optionsKey="name"
@@ -307,6 +308,7 @@ export const UICustomizations = {
           return row.status === "EXECUTION_TO_BE_DONE" ? (
             <ButtonNew
               label={t("START")}
+              title={t("START")}
               variation="primary"
               icon={"ArrowForward"}
               type="button"
@@ -319,6 +321,7 @@ export const UICustomizations = {
           ) : row.status === "RESOURCE_ESTIMATIONS_APPROVED" ? (
             <ButtonNew
               label={t("WBH_DOWNLOAD_MICROPLAN")}
+              title={t("WBH_DOWNLOAD_MICROPLAN")}
               variation="primary"
               icon={"FileDownload"}
               style={{ width: "290px" }}
@@ -329,6 +332,7 @@ export const UICustomizations = {
           ) : (
             <ButtonNew
               label={t("WBH_EDIT")}
+              title={t("WBH_EDIT")}
               variation="primary"
               icon={"Edit"}
               style={{ width: "290px" }}
@@ -596,6 +600,7 @@ export const UICustomizations = {
                 icon="ArrowForward"
                 iconFill=""
                 isSuffix
+                title={t(key)}
                 label={t(key)}
                 onClick={() => setShowPopup(true)}
                 // removed this because due to popup crashing on dev
@@ -604,7 +609,6 @@ export const UICustomizations = {
                 optionsKey=""
                 size="medium"
                 style={{}}
-                title=""
                 variation="primary"
               />
               {showPopup && (